OPT is working. Now have to integrate the LightGBM model.

+++ caches/lfo_features.h
@@ -34,8 +34,8 @@ struct LFOFeature {
 
     ~LFOFeature() {}
 
-    std::vector<uint8_t> get_vector() {
-        std::vector<uint8_t> features;
+    std::vector<double> get_vector() {
+        std::vector<double> features;
         features.push_back(size);
         features.push_back((optimizationGoal == BYTE_HIT_RATIO)? 1 : size);
         features.push_back(available_cache_size);
